Inner peace
FEAT: Began working on the projects section! I love the way it looks!
FEAT: A project can be added by simply filling out a little bit of information in a json like style. The icons will be reworked to be included in the files(I'm lazy right now)
REFACTOR: Added margins to the main header